Revert "Watchlist component fixes"
